refactor: 대피경로 안내 로직 수정
