+++ utils/mapUtil.ts
@@ -1,4 +1,5 @@
 import { FullCoord, LocationCoord } from '../types/locationCoord';
+import * as Location from 'expo-location';
 
 const EARTH_RADIUS_M = 6378137;
 
@@ -92,4 +93,20 @@ export function createCirclePolygon(
   }
 
   return coords;
+}
+
+export function getBearing(source: Location.LocationObject, dest: FullCoord) {
+  const lat1 = source.coords.latitude;
+  const lon1 = source.coords.longitude;
+  const lat2 = dest.latitude;
+  const lon2 = dest.longitude;
+
+  const y = Math.sin(lon2 - lon1) * Math.cos(lat2);
+  const x =
+    Math.cos(lat1) * Math.sin(lat2) - Math.sin(lat1) * Math.cos(lat2) * Math.cos(lon2 - lon1);
+  let bearing = Math.atan2(y, x);
+  bearing = (bearing * 180) / Math.PI;
+  bearing = (bearing + 360) % 360;
+
+  return bearing;
 }
